backend/pyproject.toml
[project]
name = "wirewan"
version = "1.0.0"
description = "WireGuard WAN Overlay Network Manager"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "greenlet>=3.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=41.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "qrcode[pil]>=7.4.0",
    "librouteros>=3.2.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.uv]
python-preference = "managed"
